First working prototype
Added color support functionally.

What the code does, simply, is takes the selected object, duplicates and hides it, and then recolors the duplicate object based on each face's vector normal to the global z axis. Essentially, it compares the global "slope" of each face to see if Mario can stand on it. Currently, the "steepness" value is hard-coded and cannot be changed without refreshing the addon, but I hope to change this shortly. The variable right now is 45 degrees, so if a face is 45 degrees or more "turned away" from the global z axis, then it's colored red. Otherwise, it's colored green.